requirements.txt
# requirements.txt

# Core Dependencies
torch==2.1.0
transformers==4.36.0
spacy==3.7.2


# NLP Models and Tools
sentencepiece==0.1.99
protobuf==4.25.1
accelerate==0.25.0

# Lightweight Models
sentence-transformers==2.2.2

# Text Processing
nltk==3.8.1
python-dateutil==2.8.2
regex==2023.10.3

# API Integration (Optional - for LLM fallback)
groq>=1.0.0
httpx>=0.27.0,<0.28.0  # Pin httpx to version compatible with groq

# Data Handling
pandas==2.1.4
numpy==1.26.2
pyyaml==6.0.1

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
colorama==0.4.6
ipykernel
ipywidgets>=8.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0